Student CRUD operations

+++ base/views.py
@@ -11,13 +11,13 @@ import json
 from rest_framework.decorators import api_view
 from django.contrib.admin.views.decorators import staff_member_required
 from accounts.models import api_key
+from django.http import JsonResponse
 
 
 @staff_member_required(login_url="/accounts/login/")
 def listEvents(request):
     events = Event.objects.all()
     return render(request, "base/eventList.html", {"events": events})
-
 
 
 def search(request):
@@ -57,6 +57,7 @@ def search1(request, event_id):
     context = {'event': event, 'event_dict': event_dict, 'count': count, 'eventss': eventss}
     return render(request, 'base/eventDetails.html', context)
 
+
 @staff_member_required
 def create_event(request):
     if request.POST.get("action") == "post":
@@ -82,13 +83,12 @@ def updateEvent(request, event_id):
     event_dict.pop("student_lists")
     if request.POST.get("action") == "post":
         data = json.loads(request.POST.get("json_sent"))
-        print(data)
         for key, value in data.items():
             setattr(event, key, value)
         event.save()
         return redirect("viewEvent", event.id)
     return render(
-        request, "base/update_event.html", {"event": event, "event_dict": event_dict}
+        request, "base/updateEvent.html", {"event": event, "event_dict": event_dict}
     )
 
 
@@ -124,18 +124,19 @@ def handle_uploaded_file(f):
 @staff_member_required
 def upload_student_list(request, event_id):
     if request.method == "POST":
-        handle_uploaded_file(request.FILES["file"])
-        data = request.FILES["file"]
+        data = request.FILES.get("file")
         type1 = request.POST.get("input")
         list1 = StudentList(type=type1)
         event = Event.objects.get(id=event_id)
-        with open("base/upload/" + data.name, "r") as csv_file:
-            datas = csv.DictReader(csv_file)
-            for row in datas:
-                stu = Student(**row)
-                stu.save()
-                list1.list.append(stu)
-        os.remove("base/upload/" + data.name)
+        if data is not None:
+            handle_uploaded_file(request.FILES["file"])
+            with open("base/upload/" + data.name, "r") as csv_file:
+                datas = csv.DictReader(csv_file)
+                for row in datas:
+                    stu = Student(**row)
+                    stu.save()
+                    list1.list.append(stu)
+            os.remove("base/upload/" + data.name)
         list1.save()
         event.student_lists.append(list1)
         event.save()
@@ -143,6 +144,45 @@ def upload_student_list(request, event_id):
     else:
         form = FileForm()
     return render(request, "base/upload_file.html", {'form': form})
+
+
+@staff_member_required
+def updateStudent(request, student_id):
+    stu = Student.objects.get(id=student_id)
+    list = StudentList.objects.filter(list__contains=stu.id)[0]
+    event = Event.objects.filter(student_lists__contains=list.id)
+    stu_dict = stu.to_mongo().to_dict()
+    stu_dict.pop("_id")
+    if request.POST.get("action") == "post":
+        data = json.loads(request.POST.get("json_sent"))
+        for key, value in data.items():
+            setattr(stu, key, value)
+        stu.save()
+        return redirect("viewStudents", list.id)
+    return render(request, "base/updateStudent.html", {
+        "stu": stu, "stu_dict": stu_dict, "event": event, "stulist": list
+    })
+
+
+@staff_member_required
+def createStudent(request, list_id):
+    list = StudentList.objects.get(id=list_id)
+    if request.POST.get("action") == "post":
+        data = json.loads(request.POST.get("json_sent"))
+        stu = Student(**data)
+        stu.save()
+        list.list.append(stu)
+        list.save()
+        return redirect("viewStudents", list_id)
+    return render(request, "base/addStudent.html", {'list': list})
+
+
+@staff_member_required
+def deleteStudent(request, student_id):
+    stu = Student.objects.get(id=student_id)
+    list = StudentList.objects.get(list__contains=stu.id)
+    stu.delete()
+    return redirect("viewStudents", list.id)
 
 
 @api_view(['POST'])
@@ -161,3 +201,9 @@ def event_APIView(request):
             return Response("INVALID API KEY", status=status.HTTP_403_FORBIDDEN)
     else:
         return Response("API KEY NOT PROVIDED", status=status.HTTP_401_UNAUTHORIZED)
+
+
+def key_viewer(request):
+    key = api_key.objects.get(user=request.user)
+    if key is not None:
+        return JsonResponse({'key': key.apiKey})
